# Optimize AS::LogSubscriber