@jhawthorn 2022-08-10
This seems to hurt the benchmark from #44469 a little bit... still faster than before that PR, but measurable. Maybe due to the extra filtering?

Before:

```
$ be ruby benchmark_notifications.rb
Warming up --------------------------------------
               timed    26.252k i/100ms
     timed_monotonic    33.989k i/100ms
        event_object    23.774k i/100ms
             evented    30.815k i/100ms
        unsubscribed   383.906k i/100ms
Calculating -------------------------------------
               timed    266.360k (± 0.2%) i/s -      1.339M in   5.026492s
     timed_monotonic    339.195k (± 0.3%) i/s -      1.699M in   5.010288s
        event_object    236.727k (± 0.6%) i/s -      1.189M in   5.021568s
             evented    308.301k (± 0.2%) i/s -      1.572M in   5.097533s
        unsubscribed      3.838M (± 0.4%) i/s -     19.195M in   5.002079s

```

After:

```

$ be ruby benchmark_notifications.rb
Warming up --------------------------------------
               timed    17.996k i/100ms
     timed_monotonic    20.696k i/100ms
        event_object    16.694k i/100ms
             evented    19.664k i/100ms
        unsubscribed   326.481k i/100ms
Calculating -------------------------------------
               timed    178.232k (± 0.4%) i/s -    899.800k in   5.048546s
     timed_monotonic    206.379k (± 0.4%) i/s -      1.035M in   5.014147s
        event_object    167.198k (± 0.3%) i/s -    851.394k in   5.092191s
             evented    197.051k (± 0.2%) i/s -      1.003M in   5.089377s
        unsubscribed      3.284M (± 0.3%) i/s -     16.651M in   5.069612s
```