@jhawthorn 2022-08-11
One option might be creating a new group_klass for these `silencable?` subscribers. They're all the same type, so it should just be one more type of group, that way we'd only pay this extra cost for the actually silencable topics and subscribers.