@casperisfine 2022-08-11
Thanks for pointing out you benchmark. I managed to get back to only ~5% slower.

Not sure we can do much better as ultimately not being able to cache the whole thing necessarily is a bit more work.

However my assumption here is that some frequent events are only subscribed by various `LogSubscribers` that end up being noops in most case. If that assumption holds up, then we're way way faster. However if they are indeed subscribed in prod, then we'd take the 5% hit.

<details>
<summary>results</summary>

```
=== timed ===
Warming up --------------------------------------
log-subscriber-silenced
                        34.802k i/100ms
Calculating -------------------------------------
log-subscriber-silenced
                        347.120k (± 0.5%) i/s -      1.740M in   5.013076s

Comparison:
                main:   363889.2 i/s
log-subscriber-silenced:   347119.9 i/s - 1.05x  (± 0.00) slower


=== timed_monotonic ===
Warming up --------------------------------------
log-subscriber-silenced
                        42.030k i/100ms
Calculating -------------------------------------
log-subscriber-silenced
                        420.788k (± 0.4%) i/s -      2.144M in   5.094174s

Comparison:
                main:   447503.6 i/s
log-subscriber-silenced:   420787.8 i/s - 1.06x  (± 0.00) slower


=== event_object ===
Warming up --------------------------------------
log-subscriber-silenced
                        31.650k i/100ms
Calculating -------------------------------------
log-subscriber-silenced
                        316.084k (± 0.6%) i/s -      1.582M in   5.006794s

Comparison:
                main:   329884.8 i/s
log-subscriber-silenced:   316083.7 i/s - 1.04x  (± 0.00) slower


=== evented ===
Warming up --------------------------------------
log-subscriber-silenced
                        37.191k i/100ms
Calculating -------------------------------------
log-subscriber-silenced
                        372.250k (± 0.4%) i/s -      1.897M in   5.095432s

Comparison:
                main:   387598.9 i/s
log-subscriber-silenced:   372250.5 i/s - 1.04x  (± 0.00) slower


=== unsubscribed ===
Warming up --------------------------------------
log-subscriber-silenced
                       405.990k i/100ms
Calculating -------------------------------------
log-subscriber-silenced
                          4.069M (± 0.3%) i/s -     20.705M in   5.089025s

Comparison:
                main:  4090502.6 i/s
log-subscriber-silenced:  4068702.3 i/s - same-ish: difference falls within error


=== silenced ===
Warming up --------------------------------------
log-subscriber-silenced
                       104.296k i/100ms
Calculating -------------------------------------
log-subscriber-silenced
                          1.046M (± 0.3%) i/s -      5.319M in   5.086327s

Comparison:
log-subscriber-silenced:  1045773.3 i/s
                main:   390009.7 i/s - 2.68x  (± 0.00) slower

```
</details>
<details>
<summary>benchmark</summary>
```ruby
# frozen_string_literal: true

require 'fileutils'
require 'benchmark/ips'
require 'active_support/all'

BRANCH = `git rev-parse --abbrev-ref HEAD`.strip
SAVE_DIR = "/tmp/bench-notifications/"
FileUtils.mkdir_p(SAVE_DIR)

class EventedSubscriber
  def start(name, id, payload); end
  def finish(name, id, payload); end
end

class SilencedSubscriber < EventedSubscriber
  def silenced?(_name)
    true
  end
end

10.times do
  ActiveSupport::Notifications.subscribe("timed") do |name, started, ended, id, payload|
  end

  ActiveSupport::Notifications.monotonic_subscribe("timed_monotonic") do |name, started, ended, id, payload|
  end

  ActiveSupport::Notifications.subscribe("event_object") do |event|
  end

  ActiveSupport::Notifications.subscribe("evented", EventedSubscriber.new)

  ActiveSupport::Notifications.subscribe("silenced", SilencedSubscriber.new)
end

def bench(name, &block)
  puts "=== #{name} ==="
  Benchmark.ips do |x|
    x.report(BRANCH, &block)
    x.save! "/tmp/bench-notifications/#{name}"
    x.compare!
  end
  puts
end

bench("timed") do
  ActiveSupport::Notifications.instrument("timed"){}
end

bench("timed_monotonic") do
  ActiveSupport::Notifications.instrument("timed_monotonic"){}
end

bench("event_object") do
  ActiveSupport::Notifications.instrument("event_object"){}
end

bench("evented") do
  ActiveSupport::Notifications.instrument("evented"){}
end

bench("unsubscribed") do
  ActiveSupport::Notifications.instrument("unsubscribed"){}
end

bench("silenced") do
  ActiveSupport::Notifications.instrument("silenced"){}
end
```
</benchmark>