@casperisfine 2022-08-12
Thanks for the review John!

One thing I realized I didn't handle is methods defined after `attach_to` was called, I'll need to do another pass.